cleaned up src directory and moved troubleshooting scripts out of the way Co-authored-by: @aidynhillyard @JAndrzGutierrez
