+++ src/App.js
@@ -1,5 +1,6 @@
-import logo from './logo.svg';
+
 import './App.css';
+
 
 function App() {
   return (
