+++ src/index.js
@@ -4,7 +4,7 @@ import { BrowserRouter as Router } from 'react-router-dom';
 import 'bootstrap/dist/css/bootstrap.min.css';
 import './index.css';
 import App from './App';
-import reportWebVitals from './reportWebVitals';
+import reportWebVitals from './troubleshooting/reportWebVitals';
 
 
 ReactDOM.render(
